RAQSAPI: A R Interface to The United States Environmental Protection Agency's Air Quality System Data Mart RESTful API server

RAQSAPI is a package for R that connects the R programming environment to the United States Environmental protection agency's Air Quality System (AQS) Data Mart API for retrieval of air monitoring data.

There are two things that you must do before using this package.

If you have not done so yet register your username with Data Mart

Every time this library is reloaded AQS_API_credentials() function must be called before continuing.

please use vignette(RAQSAPI) for more details about this package.

EPA Disclaimer: This software/application was developed by the U.S. Environmental Protection Agency (USEPA). No warranty expressed or implied is made regarding the accuracy or utility of the system, nor shall the act of distribution constitute any such warranty. The USEPA has relinquished control of the information and no longer has responsibility to protect the integrity, confidentiality or availability of the information. Any reference to specific commercial products, processes, or services by service mark, trademark, manufacturer, or otherwise, does not constitute or imply their endorsement, recommendation or favoring by the USEPA. The USEPA seal and logo shall not be used in any manner to imply endorsement of any commercial product or activity by the USEPA or the United States Government.

See also

Author

Maintainer: Clinton Mccrowey mccrowey.clinton@epa.gov (ORCID) (employer: United States Environmental Protection Agency, Region 3)

Authors:

Clinton Mccrowey mccrowey.clinton@epa.gov (ORCID) (employer: United States Environmental Protection Agency, Region 3)

United States Environmental Protection Agency (ROR) (https://www.epa.gov) [copyright holder, contributor, data contributor, funder, reviewer]

Other contributors:

Timothy Sharac (employer: United States Environmental Protection Agency) [contributor, reviewer]

Nick Mangus (employer: United States Environmental Protection Agency) [reviewer]

Doug Jager (employer: United States Environmental Protection Agency) [contributor, reviewer]

Ryan Brown (employer: United States Environmental Protection Agency, Region 4) [contributor, reviewer]

Daniel Garver (employer: United States Environmental Protection Agency, Region 4) [contributor, reviewer]

Benjamin Wells (employer: United States Environmental Protection Agency) [contributor, reviewer]

Jeffrey Hollister hollister.jeff@epa.gov (ORCID) (employer: United States Environmental Protection Agency) [reviewer]

Cynthia Stahl (ORCID) (employer: United States Environmental Protection Agency, Region 3) [reviewer]

Hayley Brittingham (employer: Neptune and Company) [reviewer, contractor]

Neptune and Company (https://www.neptuneinc.org/) [contractor, reviewer]

Edward Andrews edwards.andrews@wv.gov (employer: West Virginia Department of Environmental Protection) [reviewer, contributor]

Maria Morresi Morresi.Maria@epa.gov (employer: ASRC Federal Analytical Service) [contractor, reviewer]

ASRC Federal Analytical Service (ROR) (https://www.asrcfederal.com/) [reviewer]

Milan Malfait milan.malfait94@gmail.com (ORCID) [reviewer]

Rainer M Krug Rainer@krugs.de Rainer.Krug@sib.swiss [reviewer]

SIB Swiss Institute of Bioinformatics (ROR) (https://www.sib.swiss) [reviewer]

Mark Padgham mark.padgham@email.com (ORCID) [reviewer]